Refactor and document ast.dynamic.
