+++ nengo_spa/__init__.py
@@ -1,6 +1,6 @@
 from nengo_spa import version
 
-from nengo_spa.ast.dynamic import ActionSelection
+from nengo_spa.actions import ActionSelection
 from nengo_spa.ast.symbolic import sym
 from nengo_spa.builtins import dot, reinterpret, translate
 from nengo_spa.examine import pairs, similarity, text
