+++ nengo_spa/actions.py
@@ -1,112 +1,115 @@
-try:
-    from collections.abc import Sequence
-except ImportError:
-    from collections import Sequence
-import weakref
+from nengo.network import Network as NengoNetwork
+import numpy as np
 
-from nengo.exceptions import NetworkContextError
-from nengo.utils.compat import is_integer
-
-from nengo_spa.ast.base import ActionSet, ConstructionContext
-from nengo_spa.exceptions import SpaConstructionError
-from nengo_spa.network import Network
+from nengo_spa.ast.base import Fixed, infer_types, Node
+from nengo_spa.ast import dynamic
+from nengo_spa.types import TScalar
 
 
-class AstAccessor(Sequence):
-    """Provides access to the root AST nodes of build action rules.
+class ModuleInput(object):
+    def __init__(self, input_, type_):
+        self.input = input_
+        self.type = type_
 
-    Nodes can either be accessed by their ordinal position in the action rules
-    or by the name provided in the action rules.
-    """
-
-    def __init__(self, ast):
-        self.ast = ast
-        self.by_name = {
-            node.name: node for node in ast if hasattr(node, 'name')}
-
-    def __len__(self):
-        return len(self.ast)
-
-    def __getitem__(self, key):
-        if is_integer(key):
-            return self.ast[key]
+    def __rrshift__(self, other):
+        if not isinstance(other, Node):
+            return NotImplemented
+        if ActionSelection.active is None:
+            infer_types(self, other)
+            other.connect_to(self.input)
         else:
-            return self.by_name[key]
-
-    def __contains__(self, key):
-        return key in self.ast or key in self.by_name
-
-    def keys(self):
-        return self.by_name.keys()
-
-    def _attr_list(self, attr):
-        objs = []
-        for act_set in self.ast:
-            if hasattr(act_set, attr):
-                objs.append(getattr(act_set, attr))
-        return objs
-
-    def all_bgs(self):
-        return self._attr_list("bg")
-
-    def all_thals(self):
-        return self._attr_list("thalamus")
+            return RoutedConnection(other, self)
 
 
-class Actions(AstAccessor):
-    context = None
+class RoutedConnection(object):
+    free_floating = set()
+
+    def __init__(self, source, sink):
+        self.type = infer_types(source, sink)
+        self.source = source
+        self.sink = sink
+        RoutedConnection.free_floating.add(self)
+
+    def connect_to(self, sink):
+        return self.source.connect_to(sink)
+
+    def construct(self):
+        return self.connect_to(self.sink.input)
+
+    @property
+    def fixed(self):
+        return isinstance(self.source, Fixed)
+
+    def transform(self):
+        assert self.fixed
+        if self.type == TScalar:
+            return self.source.evaluate()
+        else:
+            return np.atleast_2d(self.source.evaluate().v).T
+
+
+class ActionSelection(object):
+    active = None
 
     def __init__(self):
-        self.blocks = []
-        self.effects = []
-        self.rules = []
-
-        super(Actions, self).__init__(self.blocks)
-
-    def add_block(self, actions, mode="bg"):
-        if len(Network.context) <= 0:
-            raise NetworkContextError(
-                "actions can only be called inside a ``with network:`` block.")
-        root_network = Network.context[-1]
-
-        if mode == "bg":
-            for i, rule in enumerate(actions):
-                rule.index = i
-            actions = [ActionSet(actions)]
-
-        for block in actions:
-            block.infer_types(None)
-        construction_context = ConstructionContext(root_network)
-        for block in actions:
-            block.construct(construction_context)
-        self.blocks += actions
-
-    def add_effect(self, effect):
-        self.effects.append(effect)
-
-    def add_rule(self, rule):
-        self.effects = [
-            e for e in self.effects if e not in rule.effects.effects]
-        self.rules.append(rule)
+        self.built = False
+        self.bg = None
+        self.thalamus = None
+        self.utilities = []
+        self.actions = []
 
     def __enter__(self):
-        if Actions.context is not None and Actions.context() is not None:
-            raise SpaConstructionError("Nesting of Action contexts is not "
-                                       "supported")
-        Actions.context = weakref.ref(self)
-
+        assert not self.built
+        if ActionSelection.active is None:
+            ActionSelection.active = self
+        else:
+            raise RuntimeError()  # FIXME better error
         return self
 
-    def __exit__(self, *args):
-        Actions.context = None
+    def __exit__(self, exc_type, exc_value, traceback):
+        ActionSelection.active = None
 
-        if args[0] is not None:
+        if exc_type is not None:
             return
 
-        if len(self.rules) > 0:
-            self.add_block(self.rules, mode="bg")
-            self.rules = []
+        if len(RoutedConnection.free_floating) > 0:
+            raise RuntimeError()  # FIXME better error
 
-        if len(self.effects) > 0:
-            self.add_block(self.effects, mode="cortical")
-            self.effects = []
+        if len(self.utilities) <= 0:
+            return
+
+        self.bg = dynamic.BasalGangliaRealization(len(self.utilities))
+        self.thalamus = dynamic.ThalamusRealization(len(self.utilities))
+        self.thalamus.connect_bg(self.bg)
+
+        for index, utility in enumerate(self.utilities):
+            self.bg.connect_input(utility.output, index=index)
+
+        for index, action in enumerate(self.actions):
+            for effect in action:
+                if effect.fixed:
+                    self.thalamus.connect_fixed(
+                        index, effect.sink.input, transform=effect.transform())
+                else:
+                    self.thalamus.construct_gate(
+                        index, net=NengoNetwork.context[-1])
+                    channel = self.thalamus.construct_channel(
+                        effect.sink.input, effect.type)
+                    effect.connect_to(channel.input)
+                    self.thalamus.connect_gate(index, channel)
+
+        self.built = True
+
+    def add_action(self, condition, *actions):
+        assert ActionSelection.active is self
+        utility = dynamic.ScalarRealization()  # FIXME should be node
+        self.utilities.append(utility)
+        self.actions.append(actions)
+        RoutedConnection.free_floating.difference_update(actions)
+        return utility
+
+
+def ifmax(condition, *actions):
+    utility = ActionSelection.active.add_action(condition, *actions)
+    condition.connect_to(utility.input)
+    return utility
